lib: Rethrow tokenizer exceptions as from the parser
If an exception is thrown by the tokenizer, the parser we catch it
and rethrow it as an instance of `parser_exception` with the same
message that the initial tokenizer one carried.

Seems resonable in order for the end-user only to care about only one
general type of parsing exception as he does not really consider the
different stages that go into forming the "AST" - or at least should not
have to :).

Signed-off-by: Kristiyan Stoimenov <kristoimenov@gmail.com>