lib+tests+app: Introduce input reading strategy
This is a big and ugly change. I won't bother splitting the commit
anyways.

Allow different input readers to be used as a strategy. The parser is
templated over it which made me move pretty much everything from
`parser.cpp` and `tokenizer.cpp` to their respective headers.

Test plan:
As of now I haven't even tested it :D. I fought hard with the compiler
in order to build it and at this point I am pretty sure that it is not
terrible. I am yet to try it out.

Note:
During this change I touched everything, so I was forced to read my code
a whole lot. Which means that I didn't really like it, so some minor
insignificant changes were made. The most important of them was
commenting - I added a bunch of it.

Signed-off-by: Kristiyan Stoimenov <kristoimenov@gmail.com>